backend/agents/assessment_creator/assessor/app/assessment_creator/model/load.py
from strands.models.bedrock import BedrockModel
import os

def load_model() -> BedrockModel:
    """Get Bedrock model client using IAM credentials."""
    return BedrockModel(
        model_id="us.anthropic.claude-haiku-4-5-20251001-v1:0",
        region_name=os.getenv("AWS_REGION", "us-west-2")
    )
